Throttler: fix race conditions in Operate() termination and in tests
