+++ go/vt/vttablet/tabletserver/throttle/throttler_test.go
@@ -331,9 +331,10 @@ func runThrottler(t *testing.T, throttler *Throttler, timeout time.Duration, f f
 	assert.True(t, throttler.IsOpen())
 	assert.False(t, throttler.IsEnabled())
 
-	ok := throttler.Enable()
+	wg := throttler.Enable()
 	defer throttler.Disable()
-	assert.True(t, ok)
+	assert.NotNil(t, wg)
+	defer wg.Wait()
 	assert.True(t, throttler.IsEnabled())
 
 	if f != nil {
@@ -380,5 +381,54 @@ func TestProbesWhileOperating(t *testing.T) {
 				}
 			}
 		})
+	})
+}
+
+// TestProbesPostDisable runs the throttler for some time, and then investigates the internal throttler maps and values.
+// While the throttler is disabled, it is technically safe to iterate those structures. However, `go test -race` disagrees,
+// which is why this test is in this *exclude_race* file
+func TestProbesPostDisable(t *testing.T) {
+	throttler := newTestThrottler()
+	runThrottler(t, throttler, 2*time.Second, nil)
+
+	time.Sleep(time.Second) // throttler's Operate() quits asynchronously. For sake of `go test -race` we allow a graceful wait.
+	probes := throttler.mysqlInventory.ClustersProbes
+	assert.NotEmpty(t, probes)
+
+	selfProbes := probes[selfStoreName]
+	t.Run("self", func(t *testing.T) {
+		assert.NotEmpty(t, selfProbes)
+		require.Equal(t, 1, len(selfProbes)) // should always be true once refreshMySQLInventory() runs
+		probe, ok := selfProbes[""]
+		assert.True(t, ok)
+		assert.NotNil(t, probe)
+
+		assert.Equal(t, "", probe.Alias)
+		assert.Nil(t, probe.Tablet)
+		assert.Equal(t, "select 1", probe.MetricQuery)
+		assert.Zero(t, atomic.LoadInt64(&probe.QueryInProgress))
+	})
+
+	shardProbes := probes[shardStoreName]
+	t.Run("shard", func(t *testing.T) {
+		assert.NotEmpty(t, shardProbes)
+		assert.Equal(t, 2, len(shardProbes)) // see fake FindAllTabletAliasesInShard above
+		for _, probe := range shardProbes {
+			require.NotNil(t, probe)
+			assert.NotEmpty(t, probe.Alias)
+			assert.NotNil(t, probe.Tablet)
+			assert.Equal(t, "select 1", probe.MetricQuery)
+			assert.Zero(t, atomic.LoadInt64(&probe.QueryInProgress))
+		}
+	})
+
+	t.Run("metrics", func(t *testing.T) {
+		assert.Equal(t, 3, len(throttler.mysqlInventory.TabletMetrics)) // 1 self tablet + 2 shard tablets
+	})
+
+	t.Run("aggregated", func(t *testing.T) {
+		assert.Zero(t, throttler.aggregatedMetrics.ItemCount()) // flushed upon Disable()
+		aggr := throttler.aggregatedMetricsSnapshot()
+		assert.Empty(t, aggr)
 	})
 }
